created new folder resources. moved readme to resources. copied text-transcript and jar to resources

+++ src/cs3500/animator/model/attributes/Color.java
@@ -198,6 +198,6 @@ public class Color {
 
   @Override
   public String toString() {
-    return "(" + r + "," + g + "," + b + ")";
+    return "(" + (int) r + "," + (int) g + "," + (int) b + ")";
   }
 }
